requirements.txt
dgl_cu113==0.9.1.post1
fair_esm==2.0.0
h5py==3.10.0
joblib==1.3.2
numpy==1.24.1
pandas==2.2.3
rdkit==2023.9.3
scikit_learn==1.3.2
scipy==1.15.1
torch==2.1.2+cu118
torch_geometric==2.6.1
